backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
langchain==0.2.1
langchain-groq==0.1.6
langchain-core==0.2.3
pydantic==2.7.1
python-dotenv==1.0.1
sse-starlette==2.1.0
httpx==0.27.0
